+++ ui/src/organizations/components/CollectorRow.tsx
@@ -17,7 +17,6 @@ import EditableName from 'src/shared/components/EditableName'
 interface Props {
   collector: Telegraf
   bucket: string
-  onDownloadConfig: (telegrafID: string, telegrafName: string) => void
   onDelete: (telegrafID: string) => void
   onUpdate: (telegraf: Telegraf) => void
   onOpenInstructions: (telegrafID: string) => void
@@ -39,12 +38,6 @@ export default class CollectorRow extends PureComponent<Props> {
           <IndexList.Cell>{bucket}</IndexList.Cell>
           <IndexList.Cell revealOnHover={true} alignment={Alignment.Right}>
             <ComponentSpacer align={Alignment.Right}>
-              <Button
-                size={ComponentSize.ExtraSmall}
-                color={ComponentColor.Secondary}
-                text={'Download Config'}
-                onClick={this.handleDownloadConfig}
-              />
               <Button
                 size={ComponentSize.ExtraSmall}
                 color={ComponentColor.Secondary}
@@ -77,13 +70,6 @@ export default class CollectorRow extends PureComponent<Props> {
 
   private handleOpenConfig = (): void => {
     this.props.onOpenTelegrafConfig(
-      this.props.collector.id,
-      this.props.collector.name
-    )
-  }
-
-  private handleDownloadConfig = (): void => {
-    this.props.onDownloadConfig(
       this.props.collector.id,
       this.props.collector.name
     )